move new header to different component and use AutomationBasicBuyAndSell feature flag
